backend/requirements.txt
matplotlib
pandas
flask
selenium
camelot-py==0.11.0
camelot-py[cv]
camelot-py[plot]
bs4
lxml
psycopg2
python-dotenv
opencv-python
PyPDF2<3.0